pyproject.toml
[project]
name = "sovereignlens"
version = "0.1.0"
description = "Sovereign AI intelligence platform"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "scrapy (>=2.14.1,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.2,<0.16.0)",
    "ipython (>=9.10.0,<10.0.0)"
]
